refactor: rename card_generator module to `card_gen`
